src/pyproject.toml
[tool.poetry]
name = "AI-bookmarker"
version = "0.1.0"
description = ""
authors = ["ktny <katsunori.t10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
langchain = "^0.0.338"
chardet = "^5.1.0"
bs4 = "^0.0.1"
lxml = "^4.9.2"
openai = "^1.3.0"
pydantic = "^1.10.9"
requests = "^2.31.0"
llama-index = "^0.8.0"
python-dotenv = "^1.0.1"
mangum = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.19.0"
aiosqlite = "^0.17.0"
httpx = "^0.23.0"
black = "^22.6.0"
flake8 = "^5.0.4"
mypy = "^0.971"
install = "^1.3.5"
isort = "^5.10.1"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
